% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.docdb_operations.R
\name{delete_db_cluster}
\alias{delete_db_cluster}
\title{Deletes a previously provisioned DB cluster}
\usage{
delete_db_cluster(DBClusterIdentifier, SkipFinalSnapshot = NULL,
  FinalDBSnapshotIdentifier = NULL)
}
\arguments{
\item{DBClusterIdentifier}{[required] The DB cluster identifier for the DB cluster to be deleted. This parameter isn't case sensitive.

Constraints:
\itemize{
\item Must match an existing \code{DBClusterIdentifier}.
}}

\item{SkipFinalSnapshot}{Determines whether a final DB cluster snapshot is created before the DB cluster is deleted. If \code{true} is specified, no DB cluster snapshot is created. If \code{false} is specified, a DB cluster snapshot is created before the DB cluster is deleted.

If \code{SkipFinalSnapshot} is \code{false}, you must specify a \code{FinalDBSnapshotIdentifier} parameter.

Default: \code{false}}

\item{FinalDBSnapshotIdentifier}{The DB cluster snapshot identifier of the new DB cluster snapshot created when \code{SkipFinalSnapshot} is set to \code{false}.

Specifying this parameter and also setting the \code{SkipFinalShapshot} parameter to \code{true} results in an error.

Constraints:
\itemize{
\item Must be from 1 to 255 letters, numbers, or hyphens.
\item The first character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}}
}
\description{
Deletes a previously provisioned DB cluster. When you delete a DB cluster, all automated backups for that DB cluster are deleted and can't be recovered. Manual DB cluster snapshots of the specified DB cluster are not deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_db_cluster(
  DBClusterIdentifier = "string",
  SkipFinalSnapshot = TRUE|FALSE,
  FinalDBSnapshotIdentifier = "string"
)
}
}

