% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.docdb_operations.R
\name{delete_db_instance}
\alias{delete_db_instance}
\title{Deletes a previously provisioned DB instance}
\usage{
delete_db_instance(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier for the DB instance to be deleted. This parameter isn't case sensitive.

Constraints:
\itemize{
\item Must match the name of an existing DB instance.
}}
}
\description{
Deletes a previously provisioned DB instance.
}
\section{Accepted Parameters}{
\preformatted{delete_db_instance(
  DBInstanceIdentifier = "string"
)
}
}

