% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.docdb_operations.R
\name{failover_db_cluster}
\alias{failover_db_cluster}
\title{Forces a failover for a DB cluster}
\usage{
failover_db_cluster(DBClusterIdentifier = NULL,
  TargetDBInstanceIdentifier = NULL)
}
\arguments{
\item{DBClusterIdentifier}{A DB cluster identifier to force a failover for. This parameter is not case sensitive.

Constraints:
\itemize{
\item Must match the identifier of an existing \code{DBCluster}.
}}

\item{TargetDBInstanceIdentifier}{The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Amazon DocumentDB replica in the DB cluster. For example, \code{mydbcluster-replica1}.}
}
\description{
Forces a failover for a DB cluster.
}
\details{
A failover for a DB cluster promotes one of the Amazon DocumentDB replicas (read-only instances) in the DB cluster to be the primary instance (the cluster writer).

If the primary instance fails, Amazon DocumentDB automatically fails over to an Amazon DocumentDB replica, if one exists. You can force a failover when you want to simulate a failure of a primary instance for testing.
}
\section{Accepted Parameters}{
\preformatted{failover_db_cluster(
  DBClusterIdentifier = "string",
  TargetDBInstanceIdentifier = "string"
)
}
}

