% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.docdb_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags on an Amazon DocumentDB resource}
\usage{
list_tags_for_resource(ResourceName, Filters = NULL)
}
\arguments{
\item{ResourceName}{[required] The Amazon DocumentDB resource with tags to be listed. This value is an Amazon Resource Name (ARN).}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Lists all tags on an Amazon DocumentDB resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  ResourceName = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}
}

