% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.docdb_package.R
\docType{package}
\name{paws.docdb-package}
\alias{paws.docdb}
\alias{paws.docdb-package}
\title{paws.docdb: Amazon DocumentDB with MongoDB compatibility}
\description{
Amazon DocumentDB API documentation
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
